+++ src/modular/bingcd/tools.rs
@@ -59,7 +59,7 @@ impl<const LIMBS: usize> Uint<LIMBS> {
     ///
     /// Executes in time variable in `length` only.
     #[inline(always)]
-    pub(super) const fn section_vartime_length<const SECTION_LIMBS: usize>(
+    pub(crate) const fn section_vartime_length<const SECTION_LIMBS: usize>(
         &self,
         idx: u32,
         length: u32,
@@ -77,7 +77,7 @@ impl<const LIMBS: usize> Uint<LIMBS> {
     ///
     /// Executes in time variable in `idx` and `length`.
     #[inline(always)]
-    pub(super) const fn section_vartime<const SECTION_LIMBS: usize>(
+    pub(crate) const fn section_vartime<const SECTION_LIMBS: usize>(
         &self,
         idx: u32,
         length: u32,
@@ -96,7 +96,7 @@ impl<const LIMBS: usize> Uint<LIMBS> {
     ///
     /// Assumes `K ≤ Uint::<SUMMARY_LIMBS>::BITS`, `n ≤ Self::BITS` and `n ≥ 2K`.
     #[inline(always)]
-    pub(super) const fn compact<const K: u32, const SUMMARY_LIMBS: usize>(
+    pub(crate) const fn compact<const K: u32, const SUMMARY_LIMBS: usize>(
         &self,
         n: u32,
     ) -> Uint<SUMMARY_LIMBS> {
